requirements.txt
dateutil
django-uuidfield
certifi==2019.9.11
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.4
django-cors-headers==3.1.0
django-model-utils==3.2.0
django-notifications-hq==1.5.0
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.10.3
djangorestframework-jsonapi==2.8.0
idna==2.7
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.1
jsonfield==2.0.2
MarkupSafe==1.1.1
openapi-codec==1.3.2
Pillow==6.1.0
psycopg2==2.8.3
PyJWT==1.7.1
pytz==2019.2
requests==2.22.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.6
